applespi: Handle command response messages correctly.
After sending any command we will receive a response message (this
includes the backlight and caps-lock commands). But that may not be the
next message received. However, all received messages are signaled by an
interrupt. So the receiving and processing of those response messages is
now left to our regular receive handler (currently all we do with the
messages is log them).